Remove vegetation plots based on completeness criteria

Remove vegetation plots based on completeness criteria

Usage

QC_completeness_criteria(
  veg_df = NULL,
  min_total_records = NULL,
  min_unique_species = NULL,
  min_perc_to_species = NULL,
  min_perc_with_abund = NULL
)

Arguments

veg_df

vegetation dataframe

min_species

(%) - sites must have >= min_species to be included

min_abund

integer - sites must have >= min_abund species with abundance data to be included (e.g., entering 90 would include all sites with 90% or more species with abundance data). WARNING: If you keep sites that are missing abundance data and intend to run vegetation analyses using abundance data, abundance values will have to be populated, otherwise that species record will be removed. If you run analyses on presence/absence data, abundance data is not needed.

Value

vegetation dataframe meeting completeness criteria

See also

Examples

QC_completeness_criteria(veg_df = my_veg_df, min_unique_species = 5, min_perc_to_species = 70, min_perc_with_abund = 70)
#> Error: object 'my_veg_df' not found