Example data set from the Sequoia and King's Canyon National Parks Soil Survey conducted by the Natural Resources Conservation Service (Soil Survey Area 792). All data is obtained from the National Soils Information System (NASIS) database.

Example data set from the Sequoia and King's Canyon National Parks Soil Survey conducted by the Natural Resources Conservation Service (Soil Survey Area 792). All data is obtained from the National Soils Information System (NASIS) database.

Usage

data(SEKI_NP_Veg)

Format

A data frame with 17444 rows and 44 variables:

siteiid

Primary key for Site table record

usiteid

Alpha-numeric code for Site

siteobsiid

Primary key for Site Observation table record

vegplotid

Alpha-numeric code for Vegetation Plot, often matching User Site ID

vegplotiid

Primary key for Vegetation Plot table record

primarydatacollector

Primary data collector for vegetation data

vegdataorigin

Origin of vegetation data (e.g., directly entered, imported, converted from other vegetation database)

ecositeid

Foreign key for Ecological Site in the Site Ecological Site History table

ecositenm

Name of Ecological Site

ecostateid

Foreign key for Ecological Site State in the Site Ecological Site History table

ecostatename

Name of Ecological Site State

commphaseid

Foreign key for Ecological Site Community Phase in the Site Ecological Site History table

commphasename

Name of Ecological Site Community Phase

cancovtotalpct

The estimated percent of the ground that is shaded by vegetation canopy at midday.

cancovtotalclass

The estimated percent of the ground that is shaded by vegetation canopy at midday, expressed as a class.

overstorycancontotalpct

The canopy cover percent of all species in the overstory stratum.

overstorycancovtotalclass

The canopy cover percent of all species in the overstory stratum, expressed as a class.

plantsym

Foreign key for Plant Symbol in the Plant table

plantsciname

The full genus and species name as listed in The PLANTS Database, USDA-NRCS, National Plant Data Center.

plantnatvernm

The most generally accepted common name of a plant.

akstratumcoverclasspct

The Alaska stratum cover class cover in percent.

speciescancovpct

The estimated canopy cover percent of a species.

speciescomppct

The percentage of a particular species in the plant, including tree, community within the plot.

speciestraceamtflag

A yes/no indicator that a 'trace' amount of a species exists. A 'trace' is defined as 'less than a measureable amount'.

vegetationstratalevel

Identifies which strata or level of the vegetation canopy the particular referenced plant is a part of.

akstratumcoverclass

General life form and horizontal layer (cover class) added to accommodate the STRATUMcv class attribute in AK SITE.

plantheightcllowerlimit

The lower height limit of the plant height class being described.

plantheightclupperlimit

The upper height limit of the plant height class being described.

planttypegroup

The designation of a plant type group being inventoried by lifeform.

horizdatnm

Datum Name.

utmzone

UTM Zone.

utmeasting

UTM Easting.

utmnorthing

UTM Northing.

latdegrees

Latitude degrees.

latminutes

Latitude minutes.

latseconds

Latitude seconds.

latdir

Latitude position north or south of the equator.

longdegrees

Longitude degrees.

longminutes

Longitude minutes.

longseconds

Longitude seconds.

longdir

Longitude east or west of Greenwich.

latstDD

Standardized latitude value in decimal degrees, in geographic coordinate system, WGS84 datum..

longstDD

Standardized longitude value in decimal degrees, in geographic coordinate system, WGS84 datum.