Big beta diversity

big_beta_diversity calculates a distance matrix using the distance measure, big beta diversity.

Usage

big_beta_diversity(
  veg_df,
  wisconsin = FALSE,
  remove_rare = FALSE,
  relative_dist = TRUE
)

Arguments

veg_df

a properly formatted vegetation dataframe (formatted_veg_df). The data should be QCed to ensure that it meets minimum data requirements (QC_vegplots)

remove_rare

decimal value defining the fraction of sites a species must be present in to be included (e.g., 0.01 means a species must be present in > 1% of plots). FALSE to include all species in dataset.

relative_dist

relativized 0-1 dissimilarity values using: $$z_i = \frac{x_i - min(x)}{max(x) - min(x)}$$

Value

a distance matrix. This distance matrix can be used for ordination or to determine the similarity between different ecological sites.

Examples

my_bbd <- big_beta_diversity(veg_df = my_veg_df)
#> Error: object 'my_veg_df' not found