Run NMDS on vegetation data

nmds_ecosite runs an NMDS using the vegan package. NMDS has numerous settings that can be adjusted, not all of which can be manipulated through the arguments of this function. This function is intended to be used as a quick way of producing ordinations. If you are interested in optimizing your NMDS, please do your own research. At a minimum, you should understand stress and evaluate the stress report in the results.

nmds_ecosite(veg_summary, ecosite, pres_abs, nmds_dim, reduce_species)

Arguments

veg_summary

veg_summary object where the ecositer::veg_summary() function was saved.

ecosite

ecosite is the ecosite(s) you want to include in an ordination. If you are interested in ordinating multiple ecosites together, concatenate the ecosites of interest (e.g., c("a", "b"))

pres_abs

use TRUE to convert data to presence/absence. This is the more robust option. If you have missing abundance values, presence/absence should be used. If you have abundance data, use FALSE.

nmds_dim

specify the number of dimensions you want in your NMDS. This value will affect stress considerably.

reduce_species

NA if you do not want to remove species. A numeric value, interpretted as percent, specifying the minimum percentage of plots a species must be present in to remain in the dataset.

Value

an NMDS object generated using the vegan package

Examples

formatted_veg <- formatted_veg_df(static_location = "C:/Users/Nathan.Roe/Documents/SEKI/CA792_veg.sqlite")
#> NOTE: some records are missing surface fragment cover
#> NOTE: some records are missing surface fragment cover
#> Warning: non-standard notation in Munsell value, use getClosestMunsellChip()
#> mixing dry colors ... [54 of 4844 horizons]
#> mixing moist colors ... [102 of 5407 horizons]
#> NOTE: some records are missing rock fragment volume
#> NOTE: all records are missing artifact volume
#> replacing missing lower horizon depths with top depth + 1cm ... [59 horizons]
#> Warning: Horizon top depths contain NA! Check depth logic with aqp::checkHzDepthLogic()
#> Warning: Horizon bottom depths contain NA! Check depth logic with aqp::checkHzDepthLogic()
#> -> QC: sites without pedons: 
#> 	Use `get('sites.missing.pedons', envir=get_soilDB_env())` for site record IDs (siteiid)
#> -> QC: duplicate pedons: 
#> 	Use `get('dup.pedon.ids', envir=get_soilDB_env())` for pedon record IDs (peiid)
#> Warning: cannot set `phiid` as unique pedon horizon key - `NA` introduced by fill=TRUE
#> -> QC: pedons missing bottom hz depths:
#> 	Use `get('missing.bottom.depths', envir=get_soilDB_env())` for pedon record IDs (peiid)
#> Joining with `by = join_by(siteiid)`
#> Joining with `by = join_by(siteiid, vegplotiid)`
my_veg_summary <- ecositer::veg_summary(veg_df = formatted_veg)
my_nmds <- ecositer::nmds_ecosite(veg_summary = my_veg_summary, ecosite = "F022AK100CA", pres_abs = TRUE, nmds_dim = 2, reduce_species = NA)
#> Error in veg_df %>% dplyr::group_by(vegplotid, plantsciname) %>% dplyr::summarise(sum(akstratumcoverclasspct)) %>%     dplyr::rename(pctcover = `sum(akstratumcoverclasspct)`): could not find function "%>%"