NMDS plot

nmds_plot produces and interactive NMDS from the object created with ecositer::nmds_ecosite. The function allows the user to define several properties, such as where to use pres/abs data, the number of dimensions, and whether to remove rare species.

nmds_plot(static_location, nmds, veg_summary)

Arguments

static_location

path to static nasis SQlite.

nmds

object created from ecositer::nmds_ecosite

veg_summary

objected created from ecositer::veg_summary

Value

interactive NMDS plot

Examples

my_nmds_plot <- nmds_plot(static_location = "C:/Users/Nathan.Roe/Documents/SEKI/CA792_veg_data.sqlite",
nmds = my_nmds, veg_summary = my_veg_summary)
#> NOTE: some records are missing surface fragment cover
#> Warning: non-standard notation in Munsell value, use getClosestMunsellChip()
#> mixing dry colors ... [54 of 4844 horizons]
#> mixing moist colors ... [102 of 5407 horizons]
#> NOTE: some records are missing rock fragment volume
#> NOTE: all records are missing artifact volume
#> replacing missing lower horizon depths with top depth + 1cm ... [59 horizons]
#> Warning: Horizon top depths contain NA! Check depth logic with aqp::checkHzDepthLogic()
#> Warning: Horizon bottom depths contain NA! Check depth logic with aqp::checkHzDepthLogic()
#> -> QC: sites without pedons: 
#> 	Use `get('sites.missing.pedons', envir=get_soilDB_env())` for site record IDs (siteiid)
#> -> QC: duplicate pedons: 
#> 	Use `get('dup.pedon.ids', envir=get_soilDB_env())` for pedon record IDs (peiid)
#> Warning: cannot set `phiid` as unique pedon horizon key - `NA` introduced by fill=TRUE
#> -> QC: pedons missing bottom hz depths:
#> 	Use `get('missing.bottom.depths', envir=get_soilDB_env())` for pedon record IDs (peiid)
#> NOTE: some records are missing surface fragment cover
#> Error in veg_df_wide_pres_abs_reduced %>% tibble::rownames_to_column("vegplot_id") %>%     dplyr::left_join(veg_data$vegplot %>% dplyr::select(vegplot_id,         site_id, akfieldecositeid) %>% dplyr::arrange(site_id,         akfieldecositeid) %>% dplyr::group_by(site_id) %>% dplyr::filter(row_number() ==         1)): could not find function "%>%"