Number of plots by ecosite

number_plots_by_ecosite describes the number of pedons and vegplots that have been recorded for each ecosite. It also lists the states/phases associated with each ecosite.

Usage

number_plots_by_site(static_location)

Arguments

static_location

the location where a static NASIS database has been saved.

Value

a dataframe describing the number of pedons and vegplots recorded for each ecosite

Examples

example_data <- access_example_data()
my_plots_by_site <- number_plots_by_site(static_location = example_data)
#> NOTE: some siteobsiid have surface fragment cover >= 100%
#> NOTE: some siteobsiid have surface fragment cover >= 100%
#> mixing dry colors ... [54 of 4847 horizons]
#> mixing moist colors ... [102 of 5407 horizons]
#> NOTE: all records are missing artifact volume
#> replacing missing lower horizon depths with top depth + 1cm ... [59 horizons]
#> Warning: Horizon top depths contain NA! Check depth logic with aqp::checkHzDepthLogic()
#> Warning: Horizon bottom depths contain NA! Check depth logic with aqp::checkHzDepthLogic()
#> -> QC: sites without pedons: 
#> 	Use `get('sites.missing.pedons', envir=get_soilDB_env())` for site record IDs (siteiid)
#> -> QC: duplicate pedons: 
#> 	Use `get('dup.pedon.ids', envir=get_soilDB_env())` for pedon record IDs (peiid)
#> -> QC: pedons with surface fragment cover >=100%: 
#> 	Use `get('surface.fragment.cover.gt100.siteobsiid', envir=get_soilDB_env())` for site observation record IDs (siteobsiid)
#> Warning: cannot set `phiid` as unique pedon horizon key - `NA` introduced by fill=TRUE
#> -> QC: pedons missing bottom hz depths:
#> 	Use `get('missing.bottom.depths', envir=get_soilDB_env())` for pedon record IDs (peiid)
#> Joining with `by = join_by(ecosite_simple)`
#> Joining with `by = join_by(ecosite_simple)`
# specify user-defined location -
# my_plots_by_site <- number_plots_by_site(static_location = "C:/Users/Nathan.Roe/Documents/SEKI/CA792_veg_data.sqlite")
head(my_plots_by_site)
#>     ecositeid ecosite_simple pedons vegplots numb_statephases
#> 1        <NA>           <NA>    145     1042                1
#> 2 F022AD102CA           D102     67       35                4
#> 3 F022AH102CA           H102     59       26                4
#> 4 F022AD104CA           D104     53       33                2
#> 5 F022AH106CA           H106     38       25                2
#> 6 F022AH101CA           H101     35       19                4
#>                                 statephases
#> 1                                        NA
#> 2 D102.1.1, D102.1.1P, D102.1.2F, D102.1.3F
#> 3 H102.1.4F, H102.1.1, H102.1.2F, H102.1.2A
#> 4                       D104.1.1, D104.1.4F
#> 5                       H106.1.1, H106.1.3F
#> 6 H101.1.2F, H101.1.1, H101.1.3F, H101.1.4F