Identify siteid(s)/vegplotid(s) in states/phases of STMs

sites_in_state_phase is a way of determining what siteids are associated with an ecosite state/phase.

sites_in_state_phase(veg_df)

Arguments

veg_df

a formatted vegetation dataframe created using formatted_veg_df

Value

a nested list of ecosites - states/phases - and the siteids associated with each state/phase

Examples

veg_df <- formatted_veg_df(static_location = "C:/Users/Nathan.Roe/Documents/SEKI/CA792_veg_data.sqlite")
#> NOTE: some records are missing surface fragment cover
#> NOTE: some records are missing surface fragment cover
#> Warning: non-standard notation in Munsell value, use getClosestMunsellChip()
#> mixing dry colors ... [54 of 4844 horizons]
#> mixing moist colors ... [102 of 5407 horizons]
#> NOTE: some records are missing rock fragment volume
#> NOTE: all records are missing artifact volume
#> replacing missing lower horizon depths with top depth + 1cm ... [59 horizons]
#> Warning: Horizon top depths contain NA! Check depth logic with aqp::checkHzDepthLogic()
#> Warning: Horizon bottom depths contain NA! Check depth logic with aqp::checkHzDepthLogic()
#> -> QC: sites without pedons: 
#> 	Use `get('sites.missing.pedons', envir=get_soilDB_env())` for site record IDs (siteiid)
#> -> QC: duplicate pedons: 
#> 	Use `get('dup.pedon.ids', envir=get_soilDB_env())` for pedon record IDs (peiid)
#> Warning: cannot set `phiid` as unique pedon horizon key - `NA` introduced by fill=TRUE
#> -> QC: pedons missing bottom hz depths:
#> 	Use `get('missing.bottom.depths', envir=get_soilDB_env())` for pedon record IDs (peiid)
#> Joining with `by = join_by(siteiid)`
#> Joining with `by = join_by(siteiid, vegplotiid)`
my_state_phase <- sites_in_state_phase(veg_df = veg_df)
head(my_state_phase)
#> $R022AB006CA
#> $R022AB006CA$B006.1.1
#>     akfieldecositeid      vegplotid
#> 1           B006.1.1  2012CA7922001
#> 22          B006.1.1  2012CA7921059
#> 41          B006.1.1  2012CA7921062
#> 60          B006.1.1  2013CA7928215
#> 76          B006.1.1  2013CA7928216
#> 93          B006.1.1  2014CA7928001
#> 109         B006.1.1  2013CA7927507
#> 124         B006.1.1  2013CA7927508
#> 141         B006.1.1  2013CA7927509
#> 155         B006.1.1  2014CA7921034
#> 169         B006.1.1  2013CA7921042
#> 192         B006.1.1  2013CA7921057
#> 209         B006.1.1  2013CA7921060
#> 222         B006.1.1  2013CA7925002
#> 235         B006.1.1  2013CA7925003
#> 241         B006.1.1  2013CA7925007
#> 257         B006.1.1  2013CA7925009
#> 272         B006.1.1  2015CA7926001
#> 288         B006.1.1 2015CA7923049N
#> 298         B006.1.1 2015CA7923054N
#> 305         B006.1.1 2015CA7923059N
#> 316         B006.1.1  2015CA7923079
#> 322         B006.1.1  2015CA7923080
#> 331         B006.1.1  2015CA7921083
#> 345         B006.1.1  2015CA7925076
#> 347         B006.1.1  2015CA7920004
#> 
#> $R022AB006CA$B006.1.3E
#>   akfieldecositeid     vegplotid
#> 1        B006.1.3E 2013CA7928214
#> 
#> $R022AB006CA$B009.1.1
#>   akfieldecositeid     vegplotid
#> 1         B009.1.1 2013CA7927511
#> 
#> $R022AB006CA$B006.1.3A
#>   akfieldecositeid     vegplotid
#> 1        B006.1.3A 2013CA7925021
#> 
#> $R022AB006CA$undefined
#>   akfieldecositeid      vegplotid
#> 1        undefined  2015CA7920006
#> 9        undefined 2016CA7925036N
#> 
#> 
#> $R022AA102CA
#> $R022AA102CA$A102.1.1
#>     akfieldecositeid      vegplotid
#> 1           A102.1.1  2012CA7922002
#> 17          A102.1.1  2012CA7929002
#> 40          A102.1.1  2013CA7927505
#> 61          A102.1.1  2013CA7921043
#> 87          A102.1.1  2015CA7925068
#> 100         A102.1.1  2015CA7925071
#> 112         A102.1.1  2015CA7923077
#> 130         A102.1.1  2015CA7923078
#> 149         A102.1.1  2015CA7921064
#> 159         A102.1.1 2015CA7921065N
#> 168         A102.1.1  2015CA7925061
#> 182         A102.1.1  2015CA7921082
#> 201         A102.1.1  2016CA7921005
#> 214         A102.1.1  2016CA7921019
#> 226         A102.1.1  2016CA7921042
#> 234         A102.1.1 2015CA7925059N
#> 250         A102.1.1  2016CA7923018
#> 
#> $R022AA102CA$A102.1.1M
#>   akfieldecositeid     vegplotid
#> 1        A102.1.1M 2013CA7927038
#> 
#> $R022AA102CA$undefined
#>    akfieldecositeid      vegplotid
#> 1         undefined  2016CA7921021
#> 7         undefined  2016CA7921023
#> 16        undefined  2016CA7925034
#> 28        undefined  S2019CA107001
#> 41        undefined  2019CA7925020
#> 47        undefined  2020CA7921004
#> 61        undefined  2020CA7925002
#> 67        undefined  2020CA7925005
#> 86        undefined  2020CA7921006
#> 98        undefined 2020CA7925007N
#> 
#> 
#> $F022AD103CA
#> $F022AD103CA$D103.1.4F
#>    akfieldecositeid     vegplotid
#> 1         D103.1.4F 2012CA7922003
#> 31        D103.1.4F 2012CA7929003
#> 
#> $F022AD103CA$D103.1.1
#>     akfieldecositeid      vegplotid
#> 1           D103.1.1  2012CA7922009
#> 42          D103.1.1  2012CA7928011
#> 71          D103.1.1  2012CA7924011
#> 93          D103.1.1  2012CA7921061
#> 115         D103.1.1  2012CA7929005
#> 132         D103.1.1  2012CA7929009
#> 151         D103.1.1  2014CA7922011
#> 175         D103.1.1  S2014CA107007
#> 187         D103.1.1  2014CA7922045
#> 195         D103.1.1  2015CA7921031
#> 229         D103.1.1 2016CA7921018N
#> 247         D103.1.1  2016CA7923044
#> 
#> $F022AD103CA$D003.1.1
#>    akfieldecositeid     vegplotid
#> 1          D003.1.1 2012CA7928004
#> 13         D003.1.1 2014CA7921064
#> 
#> $F022AD103CA$undefined
#>    akfieldecositeid      vegplotid
#> 1         undefined  2014CA7925077
#> 6         undefined 2015CA7921100N
#> 15        undefined  2016CA7921047
#> 18        undefined 2016CA7923032N
#> 29        undefined  2019CA7921018
#> 
#> $F022AD103CA$D103.1.3F
#>    akfieldecositeid     vegplotid
#> 1         D103.1.3F 2014CA7921030
#> 15        D103.1.3F 2014CA7921051
#> 41        D103.1.3F 2013CA7921033
#> 
#> $F022AD103CA$D103.1.2F
#>   akfieldecositeid     vegplotid
#> 1        D103.1.2F 2014CA7922031
#> 
#> $F022AD103CA$D003.1.3F
#>   akfieldecositeid     vegplotid
#> 1        D003.1.3F 2016CA7923046
#> 
#> 
#> $F022AD100CA
#> $F022AD100CA$D100.1.2F
#>   akfieldecositeid     vegplotid
#> 1        D100.1.2F 2012CA7922004
#> 
#> $F022AD100CA$D100.1.1
#>    akfieldecositeid     vegplotid
#> 1          D100.1.1 2012CA7922008
#> 16         D100.1.1 2012CA7929010
#> 44         D100.1.1 2013CA7929901
#> 52         D100.1.1 2013CA7929102
#> 70         D100.1.1 2013CA7929109
#> 
#> $F022AD100CA$D100.1.3F
#>    akfieldecositeid     vegplotid
#> 1         D100.1.3F 2012CA7921031
#> 16        D100.1.3F 2012CA7921065
#> 
#> $F022AD100CA$D100.1.4F
#>   akfieldecositeid     vegplotid
#> 1        D100.1.4F 2012CA7929004
#> 
#> $F022AD100CA$undefined
#>    akfieldecositeid      vegplotid
#> 1         undefined 2015CA7923082N
#> 6         undefined 2016CA7921048N
#> 8         undefined  2016CA7921050
#> 10        undefined  2016CA7923033
#> 18        undefined 2016CA7925044N
#> 23        undefined  2019CA7925014
#> 27        undefined  2019CA7925017
#> 
#> 
#> $R022AD002CA
#> $R022AD002CA$D003.1.1
#>   akfieldecositeid     vegplotid
#> 1         D003.1.1 2012CA7922005
#> 
#> $R022AD002CA$D002.1.1
#>     akfieldecositeid     vegplotid
#> 1           D002.1.1 2012CA7922010
#> 40          D002.1.1 2012CA7924004
#> 61          D002.1.1 2012CA7921070
#> 84          D002.1.1 2014CA7922034
#> 108         D002.1.1 2014CA7922038
#> 121         D002.1.1 2013CA7927501
#> 147         D002.1.1 2014CA7921060
#> 159         D002.1.1 2015CA7923095
#> 176         D002.1.1 2015CA7923099
#> 196         D002.1.1 2016CA7921009
#> 207         D002.1.1 2015CA7925084
#> 
#> $R022AD002CA$D002.1.3F
#>   akfieldecositeid     vegplotid
#> 1        D002.1.3F 2012CA7921064
#> 
#> $R022AD002CA$undefined
#>    akfieldecositeid      vegplotid
#> 1         undefined  2014CA7925056
#> 4         undefined  2014CA7921025
#> 14        undefined  2014CA7921039
#> 21        undefined 2014CA7921040N
#> 32        undefined  2013CA7920002
#> 35        undefined 2016CA7925012N
#> 39        undefined 2016CA7925016N
#> 43        undefined  2016CA7921039
#> 
#> $R022AD002CA$D001.1.1
#>   akfieldecositeid     vegplotid
#> 1         D001.1.1 2014CA7925086
#> 
#> $R022AD002CA$D007.1.1
#>    akfieldecositeid     vegplotid
#> 1          D007.1.1 2014CA7921052
#> 23         D007.1.1 2015CA7921032
#> 
#> $R022AD002CA$B004.1.1
#>   akfieldecositeid     vegplotid
#> 1         B004.1.1 2013CA7921064
#> 
#> $R022AD002CA$D004.1.1
#>   akfieldecositeid     vegplotid
#> 1         D004.1.1 2015CA7926020
#> 
#> 
#> $R022AD009CA
#> $R022AD009CA$D009.1.1
#>    akfieldecositeid     vegplotid
#> 1          D009.1.1 2012CA7922006
#> 28         D009.1.1 2013CA7927033
#> 49         D009.1.1 S2014CA107002
#> 59         D009.1.1 2014CA7921042
#> 71         D009.1.1 2014CA7921050
#> 81         D009.1.1 2014CA7922030
#> 
#> $R022AD009CA$B001.1.1
#>    akfieldecositeid      vegplotid
#> 1          B001.1.1  2012CA7928008
#> 28         B001.1.1  2012CA7929006
#> 52         B001.1.1 2015CA7921087N
#> 
#> $R022AD009CA$undefined
#>    akfieldecositeid     vegplotid
#> 1         undefined 2014CA7921041
#> 10        undefined 2017CA7921012
#> 17        undefined 2019CA7921015
#> 
#> $R022AD009CA$D004.1.1
#>   akfieldecositeid     vegplotid
#> 1         D004.1.1 2014CA7921043
#> 
#> $R022AD009CA$D103.1.2F
#>   akfieldecositeid     vegplotid
#> 1        D103.1.2F 2014CA7922025
#> 
#> $R022AD009CA$D009.1.1P
#>   akfieldecositeid     vegplotid
#> 1        D009.1.1P 2014CA7922027
#> 
#>