Summarize vegetation data by ecosite, state, and community phase

Summarize vegetation data by ecosite, state, and community phase

Usage

summarize_veg_by_ecosite(
  veg_df,
  agg_abund = TRUE,
  update_taxonomy = TRUE,
  best_vegplot = TRUE,
  use_id = TRUE
)

Arguments

veg_df

vegetation dataframe

agg_abund

logical - aggregate abundance columns using ecositer::QC_aggregate_abundance()

update_taxonomy

logical - update taxonomy from USDA PLANTS using ecositer::QC_update_taxonomy(),

best_vegplot

logical - use best vegplot when multiple for site using ecositer::QC_best_vegplot_for_site(),

use_id

logical - use state and phase id (i.e., numeric)? Alternatively, state and phase names are used

Value

list of summarized vegetation data