Create vegetation data Static NASIS

vegStaticNASIS saves a NASIS selected set to your local drive. The function utilizes soilDB::createStaticNASIS, and includes the appropriate tables to access vegetation data. The output can be used as a reliable, local copy of query results, allowing the user to change their NASIS selected set and still access the data of interest. The file extension needs to be ".sqlite"

vegStaticNASIS(output_path)

Arguments

output_path

Value

a static NASIS database that can be called independent of your current selected set in NASIS.

Details

Examples of queries to use this function with include, MLRA13_Wasilla > Pedon/Site/Transect/Vegetation Plot by usiteid (multiple) & MLRA13_Wasilla > Pedon/Site/Vegetation Plot by site area overlap table

Code written by Andrew Brown

Examples

vegStaticNASIS("C:/Users/Nathan.Roe/Documents/SEKI/CA792_veg.sqlite")
#> Loading required namespace: odbc
#> [[1]]
#> [1] TRUE
#> 
#> [[2]]
#> [1] TRUE
#> 
#> [[3]]
#> [1] TRUE
#> 
#> [[4]]
#> [1] TRUE
#> 
#> [[5]]
#> [1] TRUE
#> 
#> [[6]]
#> [1] TRUE
#> 
#> [[7]]
#> [1] TRUE
#> 
#> [[8]]
#> [1] TRUE
#> 
#> [[9]]
#> [1] TRUE
#> 
#> [[10]]
#> [1] TRUE
#> 
#> [[11]]
#> [1] TRUE
#> 
#> [[12]]
#> [1] TRUE
#> 
#> [[13]]
#> [1] TRUE
#> 
#> [[14]]
#> [1] TRUE
#> 
#> [[15]]
#> [1] TRUE
#> 
#> [[16]]
#> [1] TRUE
#> 
#> [[17]]
#> [1] TRUE
#> 
#> [[18]]
#> [1] TRUE
#> 
#> [[19]]
#> [1] TRUE
#> 
#> [[20]]
#> [1] TRUE
#> 
#> [[21]]
#> [1] TRUE
#> 
#> [[22]]
#> [1] TRUE
#> 
#> [[23]]
#> [1] TRUE
#> 
#> [[24]]
#> [1] TRUE
#> 
#> [[25]]
#> [1] TRUE
#> 
#> [[26]]
#> [1] TRUE
#> 
#> [[27]]
#> [1] TRUE
#> 
#> [[28]]
#> [1] TRUE
#> 
#> [[29]]
#> [1] TRUE
#> 
#> [[30]]
#> [1] TRUE
#> 
#> [[31]]
#> [1] TRUE
#> 
#> [[32]]
#> [1] TRUE
#> 
#> [[33]]
#> [1] TRUE
#> 
#> [[34]]
#> [1] TRUE
#> 
#> [[35]]
#> [1] TRUE
#> 
#> [[36]]
#> [1] TRUE
#> 
#> [[37]]
#> [1] TRUE
#> 
#> [[38]]
#> [1] TRUE
#> 
#> [[39]]
#> [1] TRUE
#> 
#> [[40]]
#> [1] TRUE
#> 
#> [[41]]
#> [1] TRUE
#> 
#> [[42]]
#> [1] TRUE
#> 
#> [[43]]
#> [1] TRUE
#> 
#> [[44]]
#> [1] TRUE
#> 
#> [[45]]
#> [1] TRUE
#> 
#> [[46]]
#> [1] TRUE
#> 
#> [[47]]
#> [1] TRUE
#> 
#> [[48]]
#> [1] TRUE
#> 
#> [[49]]
#> [1] TRUE
#> 
#> [[50]]
#> [1] TRUE
#> 
#> [[51]]
#> [1] TRUE
#> 
#> [[52]]
#> [1] TRUE
#> 
#> [[53]]
#> [1] TRUE
#> 
#> [[54]]
#> [1] TRUE
#> 
#> [[55]]
#> [1] TRUE
#> 
#> [[56]]
#> [1] TRUE
#> 
#> [[57]]
#> [1] TRUE
#> 
#> [[58]]
#> [1] TRUE
#> 
#> [[59]]
#> [1] TRUE
#> 
#> [[60]]
#> [1] TRUE
#> 
#> [[61]]
#> [1] TRUE
#> 
#> [[62]]
#> [1] TRUE
#> 
#> [[63]]
#> [1] TRUE
#> 
#> [[64]]
#> [1] TRUE
#> 
#> [[65]]
#> [1] TRUE
#> 
#> [[66]]
#> [1] TRUE
#> 
#> [[67]]
#> [1] TRUE
#> 
#> [[68]]
#> [1] TRUE
#> 
#> [[69]]
#> [1] TRUE
#> 
#> [[70]]
#> [1] TRUE
#> 
#> [[71]]
#> [1] TRUE
#> 
#> [[72]]
#> [1] TRUE
#> 
#> [[73]]
#> [1] TRUE
#> 
#> [[74]]
#> [1] TRUE
#> 
#> [[75]]
#> [1] TRUE
#> 
#> [[76]]
#> [1] TRUE
#> 
#> [[77]]
#> [1] TRUE
#> 
#> [[78]]
#> [1] TRUE
#> 
#> [[79]]
#> [1] TRUE
#> 
#> [[80]]
#> [1] TRUE
#> 
#> [[81]]
#> [1] TRUE
#> 
#> [[82]]
#> [1] TRUE
#> 
#> [[83]]
#> [1] TRUE
#> 
#> [[84]]
#> [1] TRUE
#> 
#> [[85]]
#> [1] TRUE
#> 
#> [[86]]
#> [1] TRUE
#> 
#> [[87]]
#> [1] TRUE
#> 
#> [[88]]
#> [1] TRUE
#> 
#> [[89]]
#> [1] TRUE
#> 
#> [[90]]
#> [1] TRUE
#> 
#> [[91]]
#> [1] TRUE
#> 
#> [[92]]
#> [1] TRUE
#> 
#> [[93]]
#> [1] TRUE
#> 
#> [[94]]
#> [1] TRUE
#> 
#> [[95]]
#> [1] TRUE
#> 
#> [[96]]
#> [1] TRUE
#> 
#> [[97]]
#> [1] TRUE
#> 
#> [[98]]
#> [1] TRUE
#> 
#> [[99]]
#> [1] TRUE
#> 
#> [[100]]
#> [1] TRUE
#> 
#> [[101]]
#> [1] TRUE
#> 
#> [[102]]
#> [1] TRUE
#> 
#> [[103]]
#> [1] TRUE
#> 
#> [[104]]
#> [1] TRUE
#> 
#> [[105]]
#> [1] TRUE
#> 
#> [[106]]
#> [1] TRUE
#> 
#> [[107]]
#> [1] TRUE
#> 
#> [[108]]
#> [1] TRUE
#> 
#> [[109]]
#> [1] TRUE
#> 
#> [[110]]
#> [1] TRUE
#> 
#> [[111]]
#> [1] TRUE
#> 
#> [[112]]
#> [1] TRUE
#> 
#> [[113]]
#> [1] TRUE
#> 
#> [[114]]
#> [1] TRUE
#> 
#> [[115]]
#> [1] TRUE
#> 
#> [[116]]
#> [1] TRUE
#> 
#> [[117]]
#> [1] TRUE
#> 
#> [[118]]
#> [1] TRUE
#> 
#> [[119]]
#> [1] TRUE
#> 
#> [[120]]
#> [1] TRUE
#> 
#> [[121]]
#> [1] TRUE
#> 
#> [[122]]
#> [1] TRUE
#> 
#> [[123]]
#> [1] TRUE
#> 
#> [[124]]
#> [1] TRUE
#> 
#> [[125]]
#> [1] TRUE
#> 
#> [[126]]
#> [1] TRUE
#> 
#> [[127]]
#> [1] TRUE
#> 
#> [[128]]
#> [1] TRUE
#> 
#> [[129]]
#> [1] TRUE
#> 
#> [[130]]
#> [1] TRUE
#> 
#> [[131]]
#> [1] TRUE
#> 
#> [[132]]
#> [1] TRUE
#> 
#> [[133]]
#> [1] TRUE
#> 
#> [[134]]
#> [1] TRUE
#> 
#> [[135]]
#> [1] TRUE
#> 
#> [[136]]
#> [1] TRUE
#> 
#> [[137]]
#> [1] TRUE
#> 
#> [[138]]
#> [1] TRUE
#> 
#> [[139]]
#> [1] TRUE
#> 
#> [[140]]
#> [1] TRUE
#> 
#> [[141]]
#> [1] TRUE
#> 
#> [[142]]
#> [1] TRUE
#> 
#> [[143]]
#> [1] TRUE
#> 
#> [[144]]
#> [1] TRUE
#> 
#> [[145]]
#> [1] TRUE
#> 
#> [[146]]
#> [1] TRUE
#> 
#> [[147]]
#> [1] TRUE
#> 
#> [[148]]
#> [1] TRUE
#> 
#> [[149]]
#> [1] TRUE
#> 
#> [[150]]
#> [1] TRUE
#> 
#> [[151]]
#> [1] TRUE
#> 
#> [[152]]
#> [1] TRUE
#> 
#> [[153]]
#> [1] TRUE
#> 
#> [[154]]
#> [1] TRUE
#> 
#> [[155]]
#> [1] TRUE
#> 
#> [[156]]
#> [1] TRUE
#> 
#> [[157]]
#> [1] TRUE
#> 
#> [[158]]
#> [1] TRUE
#> 
#> [[159]]
#> [1] TRUE
#> 
#> [[160]]
#> [1] TRUE
#> 
#> [[161]]
#> [1] TRUE
#> 
#> [[162]]
#> [1] TRUE
#> 
#> [[163]]
#> [1] TRUE
#> 
#> [[164]]
#> [1] TRUE
#> 
#> [[165]]
#> [1] TRUE
#> 
#> [[166]]
#> [1] TRUE
#> 
#> [[167]]
#> [1] TRUE
#> 
#> [[168]]
#> [1] TRUE
#> 
#> [[169]]
#> [1] TRUE
#> 
#> [[170]]
#> [1] TRUE
#> 
#> [[171]]
#> [1] TRUE
#> 
#> [[172]]
#> [1] TRUE
#> 
#> [[173]]
#> [1] TRUE
#> 
#> [[174]]
#> [1] TRUE
#> 
#> [[175]]
#> [1] TRUE
#> 
#> [[176]]
#> [1] TRUE
#> 
#> [[177]]
#> [1] TRUE
#> 
#> [[178]]
#> [1] TRUE
#> 
#> [[179]]
#> [1] TRUE
#> 
#> [[180]]
#> [1] TRUE
#> 
#> [[181]]
#> [1] TRUE
#> 
#> [[182]]
#> [1] TRUE
#> 
#> [[183]]
#> [1] TRUE
#> 
#> [[184]]
#> [1] TRUE
#> 
#> [[185]]
#> [1] TRUE
#> 
#> [[186]]
#> [1] TRUE
#> 
#> [[187]]
#> [1] TRUE
#> 
#> [[188]]
#> [1] TRUE
#> 
#> [[189]]
#> [1] TRUE
#> 
#> [[190]]
#> [1] TRUE
#> 
#> [[191]]
#> [1] TRUE
#> 
#> [[192]]
#> [1] TRUE
#> 
#> [[193]]
#> [1] TRUE
#> 
#> [[194]]
#> [1] TRUE
#> 
#> [[195]]
#> [1] TRUE
#> 
#> [[196]]
#> [1] TRUE
#> 
#> [[197]]
#> [1] TRUE
#> 
#> [[198]]
#> [1] TRUE
#> 
#> [[199]]
#> [1] TRUE
#> 
#> [[200]]
#> [1] TRUE
#> 
#> [[201]]
#> [1] TRUE
#> 
#> [[202]]
#> [1] TRUE
#> 
#> [[203]]
#> [1] TRUE
#> 
#> [[204]]
#> [1] TRUE
#> 
#> [[205]]
#> [1] TRUE
#> 
#> [[206]]
#> [1] TRUE
#> 
#> [[207]]
#> [1] TRUE
#> 
#> [[208]]
#> [1] TRUE
#> 
#> [[209]]
#> [1] TRUE
#> 
#> [[210]]
#> [1] TRUE
#> 
#> [[211]]
#> [1] TRUE
#> 
#> [[212]]
#> [1] TRUE
#> 
#> [[213]]
#> [1] TRUE
#> 
#> [[214]]
#> [1] TRUE
#> 
#> [[215]]
#> [1] TRUE
#> 
#> [[216]]
#> [1] TRUE
#> 
#> [[217]]
#> [1] TRUE
#> 
#> [[218]]
#> [1] TRUE
#> 
#> [[219]]
#> [1] TRUE
#> 
#> [[220]]
#> [1] TRUE
#> 
#> [[221]]
#> [1] TRUE
#> 
#> [[222]]
#> [1] TRUE
#> 
#> [[223]]
#> [1] TRUE
#> 
#> [[224]]
#> [1] TRUE
#> 
#> [[225]]
#> [1] TRUE
#> 
#> [[226]]
#> [1] TRUE
#> 
#> [[227]]
#> [1] TRUE
#> 
#> [[228]]
#> [1] TRUE
#> 
#> [[229]]
#> [1] TRUE
#> 
#> [[230]]
#> [1] TRUE
#> 
#> [[231]]
#> [1] TRUE
#> 
#> [[232]]
#> [1] TRUE
#> 
#> [[233]]
#> [1] TRUE
#> 
#> [[234]]
#> [1] TRUE
#> 
#> [[235]]
#> [1] TRUE
#> 
#> [[236]]
#> [1] TRUE
#> 
#> [[237]]
#> [1] TRUE
#> 
#> [[238]]
#> [1] TRUE
#> 
#> [[239]]
#> [1] TRUE
#> 
#> [[240]]
#> [1] TRUE
#> 
#> [[241]]
#> [1] TRUE
#> 
#> [[242]]
#> [1] TRUE
#> 
#> [[243]]
#> [1] TRUE
#> 
#> [[244]]
#> [1] TRUE
#> 
#> [[245]]
#> [1] TRUE
#> 
#> [[246]]
#> [1] TRUE
#> 
#> [[247]]
#> [1] TRUE
#> 
#> [[248]]
#> [1] TRUE
#> 
#> [[249]]
#> [1] TRUE
#> 
#> [[250]]
#> [1] TRUE
#> 
#> [[251]]
#> [1] TRUE
#> 
#> [[252]]
#> [1] TRUE
#> 
#> [[253]]
#> [1] TRUE
#> 
#> [[254]]
#> [1] TRUE
#> 
#> [[255]]
#> [1] TRUE
#> 
#> [[256]]
#> [1] TRUE
#> 
#> [[257]]
#> [1] TRUE
#> 
#> [[258]]
#> [1] TRUE
#> 
#> [[259]]
#> [1] TRUE
#> 
#> [[260]]
#> [1] TRUE
#> 
#> [[261]]
#> [1] TRUE
#> 
#> [[262]]
#> [1] TRUE
#> 
#> [[263]]
#> [1] TRUE
#> 
#> [[264]]
#> [1] TRUE
#> 
#> [[265]]
#> [1] TRUE
#> 
#> [[266]]
#> [1] TRUE
#> 
#> [[267]]
#> [1] TRUE
#> 
#> [[268]]
#> [1] TRUE
#> 
#> [[269]]
#> [1] TRUE
#> 
#> [[270]]
#> [1] TRUE
#> 
#> [[271]]
#> [1] TRUE
#> 
#> [[272]]
#> [1] TRUE
#> 
#> [[273]]
#> [1] TRUE
#> 
#> [[274]]
#> [1] TRUE
#> 
#> [[275]]
#> [1] TRUE
#> 
#> [[276]]
#> [1] TRUE
#> 
#> [[277]]
#> [1] TRUE
#> 
#> [[278]]
#> [1] TRUE
#> 
#> [[279]]
#> [1] TRUE
#> 
#> [[280]]
#> [1] TRUE
#> 
#> [[281]]
#> [1] TRUE
#> 
#> [[282]]
#> [1] TRUE
#> 
#> [[283]]
#> [1] TRUE
#> 
#> [[284]]
#> [1] TRUE
#> 
#> [[285]]
#> [1] TRUE
#> 
#> [[286]]
#> [1] TRUE
#>